+++ supervisely/nn/benchmark/object_detection_benchmark.py
@@ -3,7 +3,7 @@ from supervisely.nn.benchmark.cv_tasks import CVTask
 from supervisely.nn.benchmark.evaluation.object_detection_evaluator import (
     ObjectDetectionEvaluator,
 )
-from supervisely.nn.inference import SessionJSON
+from supervisely.nn.benchmark.utils import try_set_conf_auto
 
 CONF_THRES = 0.05
 
@@ -27,13 +27,3 @@ class ObjectDetectionBenchmark(BaseBenchmark):
             self.session, CONF_THRES
         ), f"Unable to set the confidence threshold to {CONF_THRES} for evalation."
         return super()._run_inference(output_project_id, batch_size, cache_project_on_agent)
-
-
-def try_set_conf_auto(session: SessionJSON, conf: float):
-    conf_names = ["conf", "confidence", "confidence_threshold"]
-    default = session.get_default_inference_settings()
-    for name in conf_names:
-        if name in default:
-            session.inference_settings[name] = conf
-            return True
-    return False
