+++ supervisely/nn/benchmark/utils.py
@@ -1,0 +1,12 @@
+from supervisely.nn.inference import SessionJSON
+
+
+def try_set_conf_auto(session: SessionJSON, conf: float):
+    conf_names = ["conf", "confidence", "confidence_threshold"]
+    default = session.get_default_inference_settings()
+    for name in conf_names:
+        if name in default:
+            session.inference_settings[name] = conf
+            return True
+    return False
+
